// utils/types/animeAnilist.ts
export type EpisodeToBeRendered = {
  id: string;
  title: string;
  number: number;
  image?: string | null;
};

export type Episode = {
  id: string;
  title: string;
  number: number;
  image?: string | null;
  img?: string | null;
};

export type EpisodeChunk = {
  startEp: number;
  endEp: number;
  episodes: EpisodeToBeRendered[];
};